# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: EHR Core API
  version: 1.0.0
  description: Core EHR API for managing patients, encounters, orders, tasks, notes, privileges, and audit logs
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
  /admin/users:
    post:
      summary: Create EHR user
      operationId: createEHRUser
      tags:
        - Admin
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEHRUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EHRUserResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    EHRUser:
      type: object
      required:
        - id
        - userId
        - email
        - firstName
        - lastName
        - isActive
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        licenseNumber:
          type: string
        specialty:
          type: string
        npi:
          type: string
        title:
          type: string
        department:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedBy:
          type: string
    CreateEHRUserRequest:
      type: object
      required:
        - userId
        - email
        - firstName
        - lastName
      properties:
        userId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        licenseNumber:
          type: string
        specialty:
          type: string
        npi:
          type: string
        title:
          type: string
        department:
          type: string
        role:
          type: string
    EHRUserResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/EHRUser'
...